bg(auth): implemented classes on signup and signin and bring them in the same file
